dbcache3 (#12)
* Prevent integer overflow in ReadVarInt.

We don't normally use ReadVarInt from untrusted inputs, but we might
 see this in the case of corruption.

This is exposed in test_bitcoin_fuzzy.

* Introduce FastRandomContext::randbool()

* Add fs.cpp/h

* Add constant for maximum stack size

* Clarify prevector::erase and avoid swap-to-clear

* update version

* Replace includes of boost/filesystem.h with fs.h

This is step one in abstracting the use of boost::filesystem.

* Replace uses of boost::filesystem with fs

Step two in abstracting away boost::filesystem.

* Use fsbridge for fopen and freopen

Abstracts away how a path is opened to a `FILE*`.

Reduces the number of places where path is converted to a string
for anything else but printing.

* torcontrol: Use fs::path instead of std::string for private key path

* Remove `namespace fs=fs`

Having these inside functions is silly and redundant now.

* Turn TryCreateDirectory() into TryCreateDirectories()

Use case: TryCreateDirectory(GetDataDir() / "blocks" / "index") would
fail if the blocks directory was not explicitly created before.

The line that did so was in a weird location and could be removed as a
result.

* Add SizeEstimate to CDBBatch

This allows estimating the in-memory size of a LevelDB batch.

* error() in disconnect for disk corruption, not inconsistency

* Introduce CHashVerifier to hash read data

This is necessary later, when we drop the nVersion field from the undo
data. At that point deserializing and reserializing the data won't
roundtrip anymore, and thus that approach can't be used to verify
checksums anymore.

With this CHashVerifier approach, we can deserialize while hashing the
exact serialized form that was used. This is both more efficient and
more correct in that case.

* Add specialization of SipHash for 256 + 32 bit data

We'll need a version of SipHash for tuples of 256 bits and 32 bits
data, when CCoinsViewCache switches from using txids to COutPoints as
keys.

* Remove/ignore tx version in utxo and undo

* Store/allow tx metadata in all undo records

Previously, transaction metadata (height, coinbase or not, and before
the previous commit also nVersion) was only stored for undo records
that correspond to the last output of a transaction being spent.

This only saves 2 bytes per undo record. Change this to storing this
information for every undo record, and stop complaining for having it
in non-last output spends. This means that undo dat written with
this patch won't be readable by older versions anymore.

* Report on-disk size in gettxoutsetinfo

* Introduce Coin, a single unspent output

* Replace CTxInUndo with Coin

The earlier CTxInUndo class now holds the same information as the Coin
class. Instead of duplicating functionality, replace CTxInUndo with a
serialization adapter for Coin.

* Optimization: Coin&& to ApplyTxInUndo

* Introduce new per-txout CCoinsViewCache functions

The new functions are:
* CCoinsViewCache::AddCoin: Add a single COutPoint/Coin pair.
* CCoinsViewCache::SpendCoin: Remove a single COutPoint.
* AddCoins: utility function that invokes CCoinsViewCache::AddCoin for
  each output in a CTransaction.
* AccessByTxid: utility function that searches for any output with
  a given txid.
* CCoinsViewCache::AccessCoin: retrieve the Coin for a COutPoint.
* CCoinsViewCache::HaveCoins: check whether a non-empty Coin exists
  for a given COutPoint.

The AddCoin and SpendCoin methods will eventually replace ModifyCoins
and ModifyNewCoins, AddCoins will replace CCoins::FromTx, and the new
AccessCoins and HaveCoins functions will replace their per-txid
counterparts.

Note that AccessCoin for now returns a copy of the Coin object. In a
later commit it will be change to returning a const reference (which
keeps working in all call sites).

* Switch from per-tx to per-txout CCoinsViewCache methods in some places

* Switch CScriptCheck to use Coin instead of CCoins

* Switch tests from ModifyCoins to AddCoin/SpendCoin

* Remove ModifyCoins/ModifyNewCoins

* Replace CCoins-based CTxMemPool::pruneSpent with isSpent

* Switch CCoinsView and chainstate db from per-txid to per-txout

* Refactor GetUTXOStats in preparation for per-COutPoint iteration

* Only pass things committed to by tx's witness hash to CScriptCheck

* Remove unused CCoins methods

* Pack Coin more tightly

* Reduce reserved memory space for flushing

As the maximum amount of data that can be pulled into the cache due to
a block validation is much lower now (at most one CCoin entry per input
and per output), reduce the conservative estimate used to determine
flushing time.

* Upgrade from per-tx database to per-txout

* Extend coins_tests

* [MOVEONLY] Move old CCoins class to txdb.cpp

It's only used for upgrading from the old database anymore.

* Merge CCoinsViewCache's GetOutputFor and AccessCoin

* Rename CCoinsCacheEntry::coins to coin

* Increase travis unit test timeout

* scripted-diff: various renames for per-utxo consistency

* Force on-the-fly compaction during pertxout upgrade

* Create random uint64_t by adding two random 32 bit integers

* Add support for std::unordered_{map,set} to memusage.h

* misc fixes